Update: - Optimized the algorithm to calculate in the CO2 and price - Added error handeling

+++ Models/PowerplantProductionPlan.cs
@@ -3,6 +3,6 @@
     public class PowerplantProductionPlan
     {
         public string Name { get; set; }
-        public Single p { get; set; }
+        public double p { get; set; }
     }
 }
